pyproject.toml
[project]
name = "dna-database"
version = "0.1.0"
description = ""
authors = [
    {name = "Jang-myoung-gyoon",email = "mgs556@naver.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pytest (>=8.3.4,<9.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.10.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
